+++ x/agent/types/codec.go
@@ -9,6 +9,7 @@ import (
 
 func RegisterCodec(cdc *codec.LegacyAmino) {
 	cdc.RegisterConcrete(&MsgRegister{}, "axon/agent/MsgRegister", nil)
+	cdc.RegisterConcrete(&MsgAddStake{}, "axon/agent/MsgAddStake", nil)
 	cdc.RegisterConcrete(&MsgUpdateAgent{}, "axon/agent/MsgUpdateAgent", nil)
 	cdc.RegisterConcrete(&MsgHeartbeat{}, "axon/agent/MsgHeartbeat", nil)
 	cdc.RegisterConcrete(&MsgDeregister{}, "axon/agent/MsgDeregister", nil)
@@ -19,6 +20,7 @@ func RegisterCodec(cdc *codec.LegacyAmino) {
 func RegisterInterfaces(registry types.InterfaceRegistry) {
 	registry.RegisterImplementations((*sdk.Msg)(nil),
 		&MsgRegister{},
+		&MsgAddStake{},
 		&MsgUpdateAgent{},
 		&MsgHeartbeat{},
 		&MsgDeregister{},